Add BattleMath unit tests and updated Game Actor test

+++ src/main/scala/model/Rules.scala
@@ -5,7 +5,7 @@ object Rules {
   val fieldWidth = 10
   val fieldHeight = 10
 
-  val ships: Set[(Int, Int)] = Set(
+  val ships: Set[(Byte, Int)] = Set(
     (4, 1),
     (3, 2),
     (2, 3),
